File restructure + minor edits
