+++ driver.py
@@ -4,12 +4,10 @@ TODOS:      Depends on what other things we will implement.
 Description:This file will mainly be responsive for calling out GUI for now.
 """
 
-
-from gui import Gui as gui
-
+import interface as inter
 
 def main():
-    g = gui()
+    g = inter.Gui()
 
 
 if __name__ == "__main__":
