--- gui/README.md
@@ -1,1 +1,0 @@
-TEMP
