+++ gui/__init__.py
@@ -1,0 +1,2 @@
+from .collapsiblepane import CollapsiblePane as cp
+from .infodisplay import InfoDisplay as infd
