+++ gui/collapsiblepane.py
